Enable select native vector intrinsics
actually allow the given ops to take vectors
add vector overload type and apply to the relevant builtins

Build lowering functions to allow vector supporting intrinsics through

Preliminary groupshared support. keep groupshared as vectors for 6.9. They are no longer represented as inidivual groupshared scalars.

adds groupshared to the test and performs the switch to CS to allow it.

Support dot product on long vecs by expanding the inrinsic into
mul/mad ops like is done with integer dot products

Since the or() and and() intrinsics did their own scalarization, the or/and operators would never be applied to full vectors. This leaves the scalarization for the scalarization pass, which will skip it for 6.9