initial commit of longvec intrinsics test
Preliminary groupshared support

Just adds groupshared to the test and performs the switch to CS to allow it. Additionally required storing output to a buffer, which was something that needed testing anyway.

keep groupshared as vectors for 6.9

They are no longer represented as inidivual groupshared scalars, but they are still retrived one single element at a time. I'm not sure we have another way to do it just yet.