Add vk::LocalSizeId Attribute
In SPIR-V, the number of thread in the group can be specificed using the
`LocalSize` execution mode. This corresponds nicely with the
`numthreads` attribute in HLSL.

However there is another way. You can use `LocalSizeId`, which uses ids
of other instructions. This allows spec constants to be provided as a
dimention on the local size id.

This PR adds a new attribute that can be used instead of the
`numthreads` attribute. It allows constant expression or spec constants
as parameters.