[SPIRV] Refactor OpExecutionModeId
The current implementation of OpExecutionModeId assumes that the
operands must be OpConstants. However, they could be the id of a
spec constant. The first step in allowing OpExecutionModeId is to modify
the internal representation of the instruction to hold general spirv
instructions.

Part of https://github.com/microsoft/DirectXShaderCompiler/issues/5960
and https://github.com/microsoft/DirectXShaderCompiler/issues/3092.