Handle coords in load helper
Load operations previously split out the addr into separate components in individual locations where they were required. This was error prone as it had a lot of places that were missing. This changes the contents of resload and atomic helpers to contain the broken down coordinates instead of the individual Value. This catches a number of new locations, but it requires the intrinsic overloads to be defined as vec1s to handle that case and it still involves a lot of fiddly places for the code that don't produce truncation warnings when longer vectors are used. Not recommended, but functional.

Fixes #7079