[Pipelines] Extend time and add PR trigger (#7071)
When the pipeline builds and tests in a windows environment, the time
limit of 90 minutes reached. There is an insufficient amount of time
available to run the pipeline in the windows environment. So, this PR
increases the limit to 120 minutes.
Here is an example of insufficient time constraints:
https://github.com/microsoft/DirectXShaderCompiler/pull/7065
Additionally, this PR explicitly adds a trigger when a PR is created, so
that the pipeline is explicitly run when that event happens. This allows
the pipeline to be run with stricter security rules.