Use UMad for dot products on uint vectors (#7059)
Dot products on uint vectors were not using the unsigned mad operations
when they were formed using the dot intrinsic though they did when it
was the effect of the mul intrinsic. While such operations aren't very
common and less common still are the cases where the unsigned variant
makes a difference, this was an internal inconsistency as well as with
the clang implementation. This corrects the problem using a dummy op for
udot similar to how other intrinsics address similar problems to pass
the unsigned information to operation lowering.

Incidentally adds unsigned notations to all the mul intrinsic entries.
This is a non-functional change as any one overload that specifies an
unsigned variant will apply for all of them, but it is less confusing
than having half of them do it and the other half not.

The test verifies the output of a few dot operations with sint, float,
and uint types for both dot and mul intrinsics

Fixes #7058

(cherry picked from commit 1fdebc418ddf88a1eea816e4f01ee86bb0c24a46)