Add error diagnostic for semantic with parameters (#6487)
A non-null parenthesized list following a semantic was mis-parsed as a
direct initializer. When the semantic actions for a direct initializer
are performed the type of the initializer is dereferenced - however as
a semantic doesn't have a type this resulted in a NULLPTR being
dereferenced which caused a crash.

As semantics don't accept parameters any parenthesized list following
a semantic is an error. We now generate an error diagnostic if a
l-paren follows a semantic.