[SPIRV] Use copy-in/copy-out for non-declaration
DXC tries to avoid the copy-in-copy-out for some parameters when it
thinks it is possible. However, the SPIR-V spec says that if a parameter
is a pointer, it must point to a memory object declarion or a pointer to
an element in a sampler or image array.

This PR enforces this. If the pointer for the parameter is not an
OpVariable (a memory object declaration), then we do not elide the
copy-in copy-out.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7095
Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/5191