[SPIRV] Stop running ReplaceInvalidOpcPass
That pass was intended to remove derivative instructions used in
non-fragment shaders. Now they are available in more places because of
new extensions. In general,
it is better practice to issue an error if they are used in an place
they are not allowed.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7086