Create new raw buffer load lowering function
Disentangles the raw buffer lowering implmentation into an isolated
function. Alters the various places that lowering took place to call
into the common function. This function will be expanded to handle other
lowering later.

When raw buffers use a templated load with a struct, they reuse the
subscript path also used for subscripted structured buffers. Such loads
with structs containing vectors or matrices will invoke the load
lowering from within this recursive call that traverses GEPs and other
users of the original call to set up correct offsets etc.

This adapts that code to use the common load lowering that enables long
vectors within structs to be correctly loaded.

Since the code expects byte address buffers, it is not (yet) adapted to
structured buffers, so those code paths are kept as they were.

This requires the ability to override the type used by the resloadhelper
explicitly, so a member is added to accomodate the matrices vector
representation that doesn't match the types of the load call.

This also requires removing the bufIdx and offset swapping that was
done, confusingly throughout the TranslateStructBufSubscriptUser code to
account for the fact that byte address buffers have to represent offsets
using the main coord parameter in favor of passing the Resource Kind
down such that the right parameter can receive the incrementation when
necessary for longer types such as matrices. This is enabled also by
adding ResKind appropriate offset calculation in the ResLoadHelper.
ResLoadHelper also gets an opcode set based on the ResKind for both
overaloads in preparation for further expansion to different resource
kinds.