Set the resource kind as u32 for boolean resources (#7116)
It was producing a validation error previously because they are
represented as u32 and i1 is not valid for output resources. This
preserves the cmp conversions for values read from the resource, but
satisfies the validation requirement. Performs some incidental
consolidation of types that are represented as u32.

Includes a couple incidental NFC fixes as well:

Moves lit filecheck tests into the correct directory. I got CodeGenHLSL
and CodeGenDXIL mixed up.

Remove dead array handing code. The array type checks were for a type
extracted from a vector. Arrays cannot be elements of vectors. This is
leftover from when the type came from more sources. The code shows as
never executed in coverage.

Related to #7079