[OMM] Add AllocateRayQuery2 DXIL Op, add validation (#7124)
This PR adds AllocateRayQuery2 as a new DXIL Op, which is intended to be
the DXIL Op that any rayquery object initialized with the 2nd optional
template argument will be lowered to. Additionally, according to the
spec here:
https://github.com/microsoft/hlsl-specs/blob/main/proposals/0024-opacity-micromaps.md
there are new validation rules that should be run against the
allocateRayQuery* DXIL Ops. This PR implements these validation rules,
and tests them.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7113

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>