NFC: Update HLSL_INTRINSIC struct for Flags and MinShaderModel fields (#7199)
HLSL_INTRINSIC will need to be updated for SM 6.9, specifically:
- to add a new flag
- to encode minimum shader model version for an availability attribute

Changing this structure is a breaking change to the internal intrinsic
table protocol, which is used for the extension mechanism.

This change separates out the breaking change with no functional changes
for simpler review and testing.

For the new flag, this change switches to using a UINT Flags field to
make flags extensible without breaking the table format. For the
version, a UINT MinShaderModel will be the encoded version format used
elsewhere: (Major << 4) | (Minor & 0xF)

Commented code for using the MinShaderModel is provided for when a
subsequent change will implement the availability attribute checks.