[OMM] Implement front end diagnostics for OMM, including on TraceRayInline, and add Availability Attributes (#7156)
This PR addresses the front end part of OMM, defining the new flags
defined in the spec, and implementing the relevant diagnostics should
the flags be incompatible. It also adds the second template argument to
the RayQuery object, which is set to have a default value of 0 if no
explicit template argument is provided.
Fixes #7145

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>