Further flatten TranslateBufLoad
Inlines both ScalarizeElements and GenerateBufLd. ScalarizeElements was
short and only used in one place. Inlining allowed for some
simpler and more legible code. GeneratBufLd got nearly entirely emptied
by moving the 64-bit conversions out of the iterative load loop because
of the observation that such conversions are only needed by typed buffer
loads, which can never be large enough to require repeated loads.

Incidentally moves the increment of the offset parameter for iterative
raw buffer loads to the top of the loop and sets the index outside it to
avoid an unwieldy conditional and index.