Always enable SPIR-V
disables the ability to disable SPIR-V. This allows intermediate opcodes
to be consistent at least across the same compiler build regardless of
build environment. This is a requirement for IR tests that depend on
these opcodes.

Given the embrace of SPIR-V and the stability the generation target has
gained, there's no reason not to do this.

Fixes #7169