Lower RayQuery constructor to allocateRayQuery2 (#7205)
This PR connects the front end change to the back end change, by taking
the existing rayquery constructor translation, and augmenting it so that
allocaterayquery2 can be emitted as an opcode if there are 2 template
arguments.
It is independent of the shader model. If 2 template args are detected,
and the 2nd template argument has a non-zero value, it just emits
allocateRayQuery2.
A test was added to make sure that when targeting shader model 6.9,
using 2 template args where the 2nd arg is non-zero in a rayquery
declaration will produce an allocateRayQuery2 opcode.

Fixes
[#7136](https://github.com/microsoft/DirectXShaderCompiler/issues/7136)