Use Wide String variants explicitly for Windows API calls (#7235)
This PR changes some code in ExecutionTests.cpp to use the wide string
variants of Windows API calls explicitly. This is because some internal
builds will get confused about which overload to resolve the
GetModuleHandle function to. By being explicit, this should eliminate
the error that an arg can't be converted to LPCWSTR.