[SPIR-V] Fix usage of indices in subfunctions
The parameters tagged with indices are linked to a builtin.
Because their layout is different between HLSL and SPIR-V, there is a
common mechanism to handle those 'stage I/O variables'.

Usually, a local variable with the correct HLSL layout is created, and
when required, the value is copied in and copied out in the entrypoint
wrapper. Then, a function-scoped pointer is passed to sub-functions.

The issue is that `indices` marks an array which is also shared across
invocations. Meaning we cannot simple copy-in/copy-out. We are only
allowed to write to the indices touched by the shader.

This required pushing the handling to the assignment expression
handling: when a value is assigned to such builtin, the layout
transformation is done, and the builtin written to.

Issue was how to find back the Builtin from an assignment: the code
assumed the ParmDecl of the entrypoint was the only way to access this
variable, but nothing prevents the user to pass this indice array to
another function.
The simple solution is to move this out of the generic map, and have
a new field which stored the SpirvVariable we created, and allow any
HLSL function to access this as soon as the HLSLIndices attribute is
found.

Signed-off-by: Nathan Gauër <brioche@google.com>