Clean up test mistakes that failed on other platforms
I don't know why the order of the extractelements varies, but it does
even for the same build depending on whether a debugger is attached.
It would seem to be an unordered container object, but there is no such.
It just generates them as it goes and never touches them again which
would suggest the opposite order in a few cases.

They are constant expressions, that's probably why they move around, but
I can't find where to make them consistent.