[SPIRV] Allow sampled type to be half for universal (#7252)
We have a check that the sample type for an image cannot be a 16-bit
float. This is true for Vulkan, but not true for general spir-v.

We modify this check to only apply when the target env is vulkan. Wew
also move the check to spirvemitter where the error handling is better.
In its current location, the compiler continue to run with an unexpected
nullptr.

Fixes #6987
Fixes #6989

---------

Co-authored-by: Cassandra Beckley <cbeckley@google.com>