[NFC] containsLongVector -> ContainsLongVector
I provided feedback during code review that this function should be
named following LLVM conventions. That feedback did not account for the
fact that SemaHLSL is otherwise consistent using CamelCase instead of
camelCase naming.

This corrects my error by renaming to match the consistent style in
SemaHLSL.h.

I've also updated the parameter naming in the source file to conform to
LLVM style since I was in the area anyways.