Allow loading DXIL shader model without module
Adds a dxilutil function to retrieve shader model from metadata. To
enable this, the metadata helper implementation is split into a few
utility functions which will continue to throw the exceptions as they
did before, but allows passes to call them without fear of crashes as a
result of that. Passes will assume the default shader model where it
can't retrieve it for one reason or another.

Changes scalarizer to try to retrieve the shader module when no module
is found. Changes dyanmicvector to array to use the utility function and
never try to recreate the hlmodule. Instead it will try the dxilmodule
and if that isn't present, it will use the metadata for shader model
only.