[Sema] Add and test new Subobject Attribute (#7258)
This PR adds and tests a new subobject attribute. It will be useful for
checking if a given decl is a subobject decl. This functionality will be
used in https://github.com/microsoft/DirectXShaderCompiler/pull/7239
We need an attribute in order to determine whether to check its
initializer for availability attributes or not.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7257