[OMM] Add D3D Flag RAYTRACING_PIPELINE_FLAG_ALLOW_OPACITY_MICROMAPS, and run d3dreflect tests (#7239)
This PR adds a new flag,
RAYTRACING_PIPELINE_FLAG_ALLOW_OPACITY_MICROMAPS, according to the spec.
It can be used with `RaytracingPipelineConfig1` subobjects. We expect
this new flag to be represented in the output.
Additionally, d3dreflect tests are run to ensure that when a rayquery
object is using the new OMM enablement flags, that
the minimum shader model target is 6.9.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7190