Rework DXIL op overload system
Add comments explaining the new system.
Eliminate bool array in favor of array of masks for up to N dimensions.
Add NumOverloadDims instead of two-mode system.

Rework TypeSlots:
- use enum, categorize basic, limit masks to used bits
- void doesn't need a type slot (NumOverloadDims == 0 instead)
- m_OverloadTypeName only contains basic type names

Handle multi-overload in FixOverloadNames; new MayHaveNonCanonicalOverload
is used to determine whether the overload name could need fixing.

Extended overload is still a distinction because of the way the overloads
must be wrapped in an unnamed StructType.
However, it does not need a bit in the overload mask.

Renamed GetVectorType to GetStructVectorType, since it's just used to get a
struct for a particular vector type, not a vector type itself.

In hctdb.py, no longer separate extended and vector overloads, just verify correctness
of the incoming string, and add default vector overloads if necessary.

In hctdb_instrhelp.py, update according to changes in hctdb.py, and eliminate
needless, problematic, outdated comment printing.