PIX: Check for existing PIX UAV in roots sigs before adding it again (#7238)
The DXR invocation counting pass calls a function to add an output UAV
twice. As part of adding the UAV, any DXIL-defined rootsigs will be
extended to include this new UAV. If the UAV already exists in the
rootsig, we should not add it again. (Doing so results in root sig that
will fail validation.)

Note: the test is not a file-check style because dxil-defined subobjects
don't get rehydrated into the DxilModule when the output of dxc.exe is
piped into the input of opt.exe, meaning that the broken case can't be
exercised.

(cherry picked from commit a13938dd6bcd08b12ef086c834c35859f050ff3f)