Allow native vector intrinsics with fixed component types
Several intrinsics that support native vectors return the same element type regardless of overload. For native vectors, these need to return a vector of size matching the overload with the fixed element type. This provides the source generation to allow it and the changes to the related lowering operations to generate the correct DXIL.

This includes some queries that return boolean information about the vector components in a boolean vector output and also the unary bit operations which operate on integers and return integers containing information about the component bits.

Fixes #7342