DxilValidation variable capitalization coding standards
Capitalize all the variables and rename a few in DxilValidation.cpp in keeping with
https://llvm.org/docs/CodingStandards.html#name-types-functions-variables-and-enumerators-properly

As this file was easily mistaken for applying to the golden rule:
https://llvm.org/docs/CodingStandards.html#name-types-functions-variables-and-enumerators-properly
it is at serious risk of receiving changes that will get hung up by
requirements to follow the LLVM coding guidelines. This brings the cases
where variable capitalization is not in line with the coding standards
to avoid such pitfalls in the future.