Add HctGen of DXIL.rst back to build without LLVM_BUILD_DOCS required
HctGen of DXIL.rst should happen on every ordinary build, and be updated with other HctGen modified files.

This isn't about building the doc, it's about updating it when definitions change in hctdb.py.

This also brings DXIL.rst up to date.
We've been missing updates to DXIL.rst for quite a while due to this issue.