Add HctGen of DXIL.rst back to build without LLVM_BUILD_DOCS required (#7346)
HctGen of DXIL.rst should happen on every ordinary build, and be updated
with other HctGen modified files.

This isn't about building the doc, it's about updating it when
definitions change in hctdb.py.

We've been missing updates to DXIL.rst for quite a while due to this
issue, introduced
[here](https://github.com/microsoft/DirectXShaderCompiler/pull/6715/files#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aR768-R770).

This also brings DXIL.rst up to date.