Revert ADO pipelines to Ubuntu 22.04 temporarily (#7365) (#7366)
DXC seems to be building inocrrectly with GCC-13 and later, which is
causing our pre-merge testing on 24.04 to fail. This will take some time
to sort out, so in the meantime I'm reverting to 22.04 on our pipelines.

(cherry picked from commit b4a3076caa92c4e9ed05761cbcd2141591fb3f89)

Co-authored-by: Chris B <cbieneman@microsoft.com>