[CoopVec] Add Linear Algebra common header with tests (#7350)
This PR introduces the linear algebra header file, and places it in a
location that is by default included in all HLSL compilation.
The builtins in the API aren't yet defined, and depend on the #7290 PR
merging first.
The tests that have been added have temporary diagnostic messages while
7290 is in progress. They will need to be updated.
Open to feedback on better / suggested error messages, or whether there
shouldn't be any sema-level validation for these errors.

Fixes
[#7304](https://github.com/microsoft/DirectXShaderCompiler/issues/7304)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>