Change default validation behavior (#7392)
This PR changes the default validation behavior to use the internal
validator by default.
If no options are specified, the internal validator will be used, and if
it fails, then compilation fails.
The external validator can still be run but must be explicitly chosen.
Specifying internal works just as before.

There is plenty of testing and infrastructure that needs to be added to
verify this change, but that needs to be added in a separate change.
This change is step 1.

Addresses https://github.com/microsoft/DirectXShaderCompiler/issues/7389