[CoopVec] This change add two CoopVec HLSL tests to the ExecutionTest framework.
ExecutionTest::CoopVec_MulAdd:

Functional verification for the Mul() and MulAdd() HLSL APIs. The driver
matrix conversion API is tested as well. These tests should be considered
as work-in-progress as this point. They include coverage primarily for SINT8,
FLOAT16, FLOAT_E4M3, and FLOAT_E5M2. The test queries the driver for all
supported configurations and runs each one, with a filtering mechanism to
limit the set of tests to the minimal feature set. The set of tests can be
further filtered by the following TE parameters:

  CoopVecMatrixInterp: SINT8, FLOAT16, FLOAT_E4M3, ...
  CoopVecMatrixLayout: ROW_MAJOR, COLUMN_MAJOR, MUL_OPTIMAL, OUTER_PRODUCT_OPTIMAL
  CoopVecBiasInterp: SINT32, FLOAT16, FLOAT_E4M3, ...
  CoopVecInputInterp: SINT8, FLOAT16, FLOAT_E4M3, ...
  CoopVecInputType: SINT8, UINT8, SINT16, UINT16, SINT32, UINT32, FLOAT16, FLOAT32, ...
  CoopVecOutputType: SINT32, UINT32, FLOAT16, FLOAT32, ...

Filter example:
  $ TE.exe ... -p:CoopVecMatrixInterp=FLOAT16 -p:CoopVecMatrixLayout=MUL_OPTIMAL

Precision coverage is minimal at this point, using an all-ones input matrix and
test vector with ones in the first two components. This is enough to test basic
functionality, but more comprehensive tests are needed.

ExecutionTest::CoopVec_OuterProduct:

Functional verification for the OuterProductAccumulate() HLSL API. This test queries
the driver for all supported configurations and runs each one. No filtering is
currently implemented.