Add missing diagnostic argument
Two instances of the err_integer_literal_too_large diagnostic in HLSL
specific code within Sema::ActOnNumericConstant() had a missing argument.
When these diagnostics were raised this caused an assert in an assert
enabled DXC, and random corruption of the diagnostic text in a non-assert
enabled DXC.

The trivial fix is to supply the required argument.

Fixes #7425