[SER] Transform RayDesc in ScalarReplHLSL for robust lowering
ScalarReplHLSL will refuse to faltten RayDesc if some users are
incompatible. This is at odds with HLOperationLower, which always
expects a specific flattened form.
This implements explicit RayDesc flattening in ScalarReplHLSL to make this more reliable. One issue with RayDesc-from-cbuffer in RayQuery remains and is documented with a test.

Implement ScalarReplHLSL handling for RayDesc param in HL ops:
- HitObject::MakeMiss [RewriteCallArg]
- HitObject::TraceRay [RewriteCallArg]
- TraceRay            [RewriteCallArg]
- RayQuery::TraceRayInline [unchanged - flatten to RayDesc elements]

In the process:
- Fix copyout code in RewriteCallArg (ScalarReplHLSL)
- Remove obsolete tests that expected a RayDesc alloca (where RayDesc is
  now SROA'd entirely)
- Document RayQuery lowering bug when RayDesc is taken directly from
  cbuffer.

Closes #7434  Unflattened RayDesc breaking HL->DXIL lowering #7434