Refactor udt intrinsic arg copy to before SROA
Intrinsics that take UDT arguments need copy-in/copy-out.
Other aggregate args are flattened for intrinsic calls.
Previously, these operations were intermingled, driven by SROA on alloca/GV values.

There were RayDesc arguments that weren't treated consistently, leading to problems.
They should be flattened into the intrinsic arguments, but TraceRay calls didn't do this.
Doing this for TraceRay would cause an issue where the subsequent Payload UDT arg could be in a different position depending on whether it was processed before or after flattening the RayDesc.

This change is in preparation for flattening RayDesc args.
It separates the UDT copy-in/copy-out into a separate operation before SROA.

When doing the copy-in/copy-out separately, we don't expand the memcpy operations until they are handled separately during SROA.
This causes a change in some IR shape, which had to be adjusted for in some tests.
There are a couple remaining PIX pass tests that still need adjustment.
These will likely need further changes when flattening RayDesc, so maybe they should be revisited after that.

TODO:
- Fix the remaining tests
- Flatten the RayDesc parameters
- Adjust tests for accordingly