[SPIRV] Treat vk::Spirv*Type as opaque when reconstructing
It is possible to have two struct types in spir-v that are the same
except for the decorations. Sometimes we have to reconstruct the value
from one type to another.

In the case of a vk::SpirvType, we do not know anything about the type,
so this should not happen. When trying to reconstuct the value, we
should simply return the original value.

Fixes #6963