Fix -fvk-invert-y (#7447)
https://github.com/microsoft/DirectXShaderCompiler/issues/7446
This fixes some outdated documentation as well as a compile error when
enabling fvk-invert-y on lib files and makes sure that it only gets
enabled on SV_POSITION that is used in VS/GS/DS/MS (so PS doesn't get
caught in the crossfire).
Also tested the dx-position-w one and that one already has correct
behavior here.

---------

Co-authored-by: NielsbishereAlt <nb@osomi.net>