[CoopVec] Check for minimum support set (#7457)
This change refactors the CoopVec test config selection to enforce
running the minimum support set. Drivers that do not report all required
data formats and interpretations will fail the test.

The handling of `D3D12EnableExperimentalFeatures` is also refactored to
apply only to the CoopVec tests, without the use of
`-p:CooperativeVectorExperimental=*`.